Rebooting somewhat; interim checkin
