+++ src/main/java/org/microbean/interceptor/LowLevelOperation.java
@@ -1,0 +1,34 @@
+/* -*- mode: Java; c-basic-offset: 2; indent-tabs-mode: nil; coding: utf-8-unix -*-
+ *
+ * Copyright © 2023 microBean™.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+package org.microbean.interceptor;
+
+@FunctionalInterface
+interface LowLevelOperation<T> {
+
+  T invoke() throws Throwable;
+
+  static <T> T invokeUnchecked(final LowLevelOperation<? extends T> c) {
+    try {
+      return c.invoke();
+    } catch (final RuntimeException | Error e) {
+      throw e;
+    } catch (final InterruptedException e) {
+      Thread.currentThread().interrupt();
+      throw new IllegalStateException(e.getMessage(), e);
+    } catch (final Throwable e) {
+      throw new IllegalStateException(e.getMessage(), e);
+    }
+  }
+
+}
